requirements.txt
streamlit
pandas
numpy
yfinance
matplotlib
python-dotenv
feedparser
pydantic
langgraph
langchain
langchain-openai
tavily-python
